Uncomment constant definitions and add proper inline documentation.
These constants are typically things we want controlled in child themes. Instead of including the constant definitions commented out in the child starter theme, let's uncomment them and specify the default Responsive Framework values. This allows us to add proper inline documentation educating developers and designers about these constants.

It also requires a deliberate action to take place to allow these options to be changeable in the Customizer, which is usually not the case in custom child themes.