Move Foundation to regular dependencies
This is a minor commit for semantics.

Foundation should belong in “dependencies” because we will want to include these tools in our resulting build.

Dev dependenices are for tooling and packages used to help develop the build. Dependencies are for running the code.

Since Foundation is used to help our child theme run properly, it is now a dependency rather than a devDependency.