Add babel to transpile es6 to es5
This adds babel via the babelify package, and configures it using the preset-env and wordpress presets.

Since we are using Browserify to bundle scripts together, the babelify package is used as a “transform’, and applies the two presets (smart defaults `@babel/preset-env, and the wordpress recommended preset). This configuration is defined in the Gruntfile.

One thing worth noting is that the wordpress preset should already include preset-env, however I was getting errors testing this workflow pulling in Foundation scripts unless I specified it as a preset so I am keeping both in our project.