Add browserify-shim package
Makes CommonJS-incompatible files Browserifyable.

We use browserify to allow us to import the global variable attached to the window for jQuery, since we will not be bundling our js with jQuery in our build process. We expect jQuery to already be loaded on the front-end in the WordPress context, so browserify-shim allows us to import that library into our modules by polyfilling the global variable as a CommonJS module. See configuration in package.json.