Bring in a module from Foundation
Brings in the Toggle component from Foundation.

Note that this only works when browserify-shim exists. This could be due to a variety of reasons — such as the Toggle component not being written properly to work with CommonJS modules; or that it was built using browserify-shim since it used jQuery, so browserify-shim must be used in this project as well? More research will be done for this.